+++ packages/core-payment/plugins/worldline-saferpay/worldline-saferpay-client/index.ts
@@ -1,0 +1,8 @@
+export { SaferpayClient } from './api/SaferpayClient';
+
+export {
+  PaymentPageInitializeInput,
+  PaymentPageInitializeResponse,
+} from './models/PaymentPageInitialize';
+export { PaymentPageAssertInput, PaymentPageAssertResponse } from './models/PaymentPageAssert';
+export { TransactionCancelInput, TransactionCancelResponse } from './models/TransactionCancel';