+++ packages/core-payment/plugins/worldline-saferpay/worldline-saferpay-client/models/Transaction.ts
@@ -1,0 +1,6 @@
+export type TransactionStatus = 'AUTHORIZED' | 'CANCELED' | 'CAPTURED' | 'PENDING';
+
+export type TransactionAmount = {
+  Value: string;
+  CurrencyCode: string;
+};